fix: remove unused codes in web3profile
