+++ packages/plugins/Web3Profile/src/SiteAdaptor/hooks/useControlSocialPopup.ts
@@ -1,9 +1,9 @@
 import { useEffect, useRef, useState, type RefObject } from 'react'
-import { emitter } from '../../emitter.js'
+import { emitter } from '../emitter.js'
 
 const LEAVE_DURATION = 500
-export function useControlFarcasterPopup(holderRef: RefObject<HTMLDivElement | null>) {
-    const isHoverRef = useRef(false)
+export function useControlSocialPopup(holderRef: RefObject<HTMLDivElement | null>) {
+    const hoverRef = useRef(false)
     const closeTimerRef = useRef<ReturnType<typeof setTimeout>>(undefined)
 
     const [active, setActive] = useState(false)
@@ -14,14 +14,14 @@ export function useControlFarcasterPopup(holderRef: RefObject<HTMLDivElement | n
         if (!holder) return
 
         const enter = () => {
-            isHoverRef.current = true
+            hoverRef.current = true
             clearTimeout(closeTimerRef.current)
         }
         const leave = () => {
-            isHoverRef.current = false
+            hoverRef.current = false
             clearTimeout(closeTimerRef.current)
             closeTimerRef.current = setTimeout(() => {
-                if (isHoverRef.current) return
+                if (hoverRef.current) return
                 setActive(false)
             }, LEAVE_DURATION)
         }
@@ -34,7 +34,7 @@ export function useControlFarcasterPopup(holderRef: RefObject<HTMLDivElement | n
     }, [holderRef.current])
 
     useEffect(() => {
-        const unsubscribe = emitter.on('open-farcaster', () => {
+        const unsubscribe = emitter.on('open', () => {
             clearTimeout(closeTimerRef.current)
             setActive(true)
         })
@@ -45,7 +45,7 @@ export function useControlFarcasterPopup(holderRef: RefObject<HTMLDivElement | n
         const onClick = (event: MouseEvent) => {
             // `NODE.contains(other)` doesn't work for cross multiple layer of Shadow DOM
             if (event.composedPath()?.includes(holderRef.current!)) return
-            isHoverRef.current = false
+            hoverRef.current = false
             setActive(false)
         }
         document.body.addEventListener('click', onClick, true)
